load_forecast_new/requirements.txt
# SIH 2024 Load Forecasting System Requirements

# Core Framework
Flask>=2.0.0
flask-cors>=4.0.0
Werkzeug>=2.3.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Machine Learning
tensorflow>=2.10.0,<2.17.0
keras>=2.8.0

# Database & ORM
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0
supabase>=2.0.0

# API & Web
requests>=2.28.0
python-dotenv>=0.19.0
plotly>=6.0.0
dash>=3.0.0
dash-bootstrap-components>=2.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.0
click>=8.0.0
joblib>=1.0.0

# Weather Integration
beautifulsoup4>=4.11.0
lxml>=4.9.0
urllib3>=1.26.0

# Phase 1.1 - Historical Data Migration
html5lib>=1.1
selenium>=4.15.0
webdriver-manager>=4.0.1

# Testing
pytest>=7.0.0
pytest-mock>=3.0.0

# Development
streamlit>=1.40.0

# Supabase SDK (for cloud database)
supabase>=2.0.0

# Additional ML libraries
xgboost>=1.6.0
lightgbm>=3.3.0

# Time series analysis
statsmodels>=0.14.0

# Deployment
gunicorn>=20.1.0
waitress>=2.1.0

# Monitoring & Logging
python-json-logger>=2.0.0

# Security
cryptography>=41.0.0

# Data visualization (if needed for dashboard)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Additional utilities
lxml>=4.6.0
html5lib>=1.1

# Development and testing
pytest>=6.0.0
pytest-mock>=3.6.0